requirements.txt
streamlit==1.28.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
pillow==10.0.0
openai==0.28.0